+++ src/ActivityImporter.Engine/Graph/O365UsageReports/ReportLoaders/SharePointUserActivityLoader.cs
@@ -4,18 +4,17 @@ using Entities.DB.Entities.UsageReports;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 
-namespace ActivityImporter.Engine.Graph.O365UsageReports.ReportLoaders.ActivityLoaders;
+namespace ActivityImporter.Engine.Graph.O365UsageReports.ReportLoaders;
 
 // https://docs.microsoft.com/en-us/graph/api/reportroot-getonedriveactivityuserdetail?view=graph-rest-beta
-public class OneDriveUserActivityLoader : AbstractActivityLoader<OneDriveUserActivityLog, OneDriveUserActivityRecord>
+public class SharePointUserActivityLoader : AbstractActivityLoader<SharePointUserActivityLog, SharePointUserActivityRecord>
 {
-    public OneDriveUserActivityLoader(ManualGraphCallClient client, ILogger logger)
+    public SharePointUserActivityLoader(ManualGraphCallClient client, ILogger logger)
         : base(client, logger)
     {
     }
 
-    public override string ReportGraphURL => "https://graph.microsoft.com/beta/reports/getOneDriveActivityUserDetail";
-    protected override void PopulateReportSpecificMetadata(OneDriveUserActivityLog todaysLog, OneDriveUserActivityRecord userActivityReportPage)
+    protected override void PopulateReportSpecificMetadata(SharePointUserActivityLog todaysLog, SharePointUserActivityRecord userActivityReportPage)
     {
         todaysLog.SharedInternally = userActivityReportPage.SharedInternally;
         todaysLog.SharedExternally = userActivityReportPage.SharedExternally;
@@ -24,7 +23,7 @@ public class OneDriveUserActivityLoader : AbstractActivityLoader<OneDriveUserAct
         todaysLog.LastActivityDate = userActivityReportPage.LastActivityDate;
     }
 
-    protected override long CountActivity(OneDriveUserActivityRecord activityPage)
+    protected override long CountActivity(SharePointUserActivityRecord activityPage)
     {
         if (activityPage is null)
         {
@@ -40,5 +39,7 @@ public class OneDriveUserActivityLoader : AbstractActivityLoader<OneDriveUserAct
 
         return count;
     }
-    public override DbSet<OneDriveUserActivityLog> GetTable(DataContext context) => context.OneDriveUserActivityLogs;
+    public override string ReportGraphURL => "https://graph.microsoft.com/beta/reports/getSharePointActivityUserDetail";
+
+    public override DbSet<SharePointUserActivityLog> GetTable(DataContext context) => context.SharePointUserActivityLogs;
 }
